wire v0x03: source_id + target_group widened to u16 LE (#40)
Bumps kProtocolVersion 0x02 -> 0x03. Header source_id and every
payload's target_group widen from 1 byte to 2 bytes little-endian.

Motivation is futureproofing:
- source_id 191 slots -> 65534 slots. Birthday-paradox says 50%
  collision at ~14 concurrent Directors on v2 (191 slots), enough
  to hit at a single dense personal-rig stage. v3 pushes that
  ceiling well past any realistic crowd.
- target_group 255 slots -> 65534 slots. Stadium-scale seat-level
  addressing headroom - not needed by current fx but locks in the
  wire so a future addressing scheme doesn't need a second break.

DAL / config / NVS all stay at u8. UI unchanged. The u16 width is
purely wire-side; top byte is zero for current-generation Directors,
and a later NVS/UI widening slots in without another wire break.

Header layout (v3): magic(2) + version(1) + source_id(2 LE) + seq(1)
+ hop(1) + type(1) + payload_len(1) = 9 bytes (v2 was 8).
kHopCountOffset 5 -> 6. Every LIGHT_* / TEXT / BITMAP / CLEAR payload
grows +1 byte. LIGHT_PULSE 9 -> 10, WASH 16 -> 17, TEXT prefix 6 -> 7,
BITMAP_HEADER 37 -> 38, BITMAP_PLANE prefix 5 -> 6, CLEAR 3 -> 4.

Old-firmware devices reject v3 at the version check; this is a hard
cutover, not a compatible extension. Coordinated with the tildagon-
firmware v0x03 PR - both merge in one bench session to avoid mixed-
version network state.

Test coverage: 36/36 native_espnow (existing byte-for-byte fixtures
regenerated + new v3 wide-range coverage), 11/11 native_modes,
56/56 native_dal, 14/14 native_effects, 15/15 native_audio. All four
device firmwares build clean (Plus2, S3, Atom Lite, AtomS3R-PoE).

Tildagon firmware + docs land in separate PRs on the same cutover.

+++ include/transport/espnow/tofu_lock.h
@@ -68,7 +68,7 @@ public:
     //     NOT update the liveness timer (they're orchestrator-
     //     origin content, not Director identity).
     bool admit(MessageType msg_type,
-               uint8_t     source_id,
+               uint16_t    source_id,          // v3: widened to u16
                uint8_t     channel,
                uint32_t    now_ms);
 
@@ -85,12 +85,12 @@ public:
     void clear();
 
     // Observers.
-    bool    is_locked() const { return has_lock_; }
-    uint8_t locked_id() const { return locked_id_; }   // undefined when !is_locked()
+    bool     is_locked() const { return has_lock_; }
+    uint16_t locked_id() const { return locked_id_; }   // undefined when !is_locked()
 
 private:
     bool     has_lock_       = false;
-    uint8_t  locked_id_      = 0;
+    uint16_t locked_id_      = 0;                       // v3: widened to u16
     uint32_t last_frame_ms_  = 0;
 };
 
@@ -106,7 +106,7 @@ private:
 // Returns the number of characters written (excluding the trailing
 // NUL), or 0 if nothing was written.
 size_t format_tofu_lock_label(bool     is_locked,
-                              uint8_t  locked_id_value,
+                              uint16_t locked_id_value,   // v3: widened to u16
                               char*    buf,
                               size_t   buflen);
